Rename data product -> subarray product
This is done to reduce confusion of the SP concept of a product (e.g.
'array_1_c856M4k') with the CBF data product itself (e.g. 'c856M4k').
The SP product actually consists of the combination of the subarray and
CBF data product, which uniquely identifies it.

The KATCP request 'data-product-configure' is the only holdout as this
would require a change to the ICD and who wants that?