Clean up requirements handling
All package dependencies are installed via requirements.txt, after which this
package may be installed in the case of Docker, or not in the case of Travis.

Remove all vestiges of install_requires in setup.py, as well as the now defunct
test-requirements.txt. The latter was created for CAM's benefit but turns out
to be more of a hassle.