Added a GraphResolver class to allow graph choice overrides
The principle use for this class is in specifying command line
overrides using --graph-override=<subarray_product_id>:<graph_name>

This will prevent the default behaviour of using the rightmost
underscore led text as the graph name and instead use the user
specified graph.

The current use case is to allow selection of different beamformer
backends based on which subarray is used. e.g.:

   --graph-override=array_1_bc856M4k:bec856M4k

which will use the engineering beamformer backend when subarray_1
is used for beamformer observations.