Increase capture-done timeout
This is need when the array is deconfigured immediately after a
capture-done is issued. It takes some time to serialise the
telescope state into the HDF5 file, and capture-done is likely
to timeout on long runs (3hours +).

Needs to be revisited at some stage to remove all these damn
timeouts...