Record dependent images in the Docker image metadata
This adds a za.ac.kat.sdp.image-depends label to the image, which
contains a space-separated list of images that form part of the graph.

With this change, builds must be done by running scripts/docker_build.sh
instead of docker build. If one forgets and runs docker build, you get a
helpful error message.