Renaming and refactoring
Node becomes LogicalTask, Task becomes PhysicalTask. This allows for
other types of nodes in the graph in future (external services and
multicast groups).

More of the logic is pushed into the physical classes, rather than
having logical classes be aware of the physical ones and having cyclic
dependencies between them.