GPU support, plus misc other
See the new section in the docstring for scheduler.py for a description
of how the GPU support works. There is also a little preparatory work
for handling more involved network interface requests and for NUMA
awareness, but it is not fully implemented. However, new schemas are
defined for passing the attributes, adhering to the limitations on the
character set imposed by Mesos.