Cycle amongst ports rather than reusing old ones
This reduces the chance of accidentally talking to the wrong process
when it dies and is replaced by another one on the same port.

See https://issues.apache.org/jira/browse/MESOS-2802 for discussions
about alternative approaches to this issue. Nothing has been done in
Mesos itself yet.

Fixes SR-704.