Add script to reverse-proxy signal displays
This adds a new script, haproxy_disp.py, that dynamically configures an
haproxy install to map paths on a frontend to the different signal
display servers. This will allow for stable URLs for the signal
displays.

It's not perfect because the data port isn't proxied; the browser
connects to it directly. This means that the page needs to be refreshed
after a subarray is taken down and brought back up.