Run bf_ingest.py with realtime scheduling
This should hopefully reduce the amount of time the HDF5 beamformer
capture gets scheduled out. It makes use of the schedrr tool I stuck in
the base image (ska-sa/katsdpdockerbase#3).

This should eventually be added to other services that struggle to get
their scheduling due, particularly PTUSE ingest, but this is still
experimental. PTUSE ingest may also want to use it only on particular
subprocesses, so that low-priority processes avoid getting pre-empted.