Increase memory allocation to telstate (redis)
A recent 12 hour observation showed substantial memory pressure
on redis, which led to it being very unresponsive and timing out
on requests. Docker stats showed it was using pretty much the whole
allocation.

This is a temporary fix to get the telescope working again, a
more detailed review of memory needed will be done