Do not wait on timeplot's data_port
This is to handle the transition to timeplot no longer having a data
port. Rather than remove the data port completely (which would break
things if we needed to roll back to an older version of timeplot), we
pass the command-line option but do not wait for that port to open.