Connect to cam2telstate before starting other services
This is to address SR-860 (waiting for cam2telstate to be ready before
data-product-configure returns).

This depends on the cam2telstate-no-wait branch of katsdpingest and the
early-telstate branch of katcbfsim.

The ?capture-meta call to the simulator is dropped, because it used to
be needed to get telstate populated, but the early-telstate branch
populates it on startup.