Support interposing an arbitrary wrapper script
This allows an arbitrary script (fetched by URL) to be injected into
each container and run instead of the "real" script. This is intended
for development purposes only (my current use case is to run cal with a
profiler, but the approach is flexible e.g. it could grab code from the
developer's machine and run that instead of the version in the
container.