Update CPU and memory usage calculation for cal
This gives it more CPUs to account for it now using multi-threading. The
scaling is changed from linear in antennas to linear in baselines, based
on what I've seen of the code while optimising. I've allows more slots
for intermediate results because cal has been changed to use coarser
chunking in time.