Allow wrapper script to be in an archive
If the upload is a tarball, it is expected to contain a file called
`wrapper` which will be run.

The user override to root is removed. It was done only to allow writing
to the sandbox, but with an archive it is possible to include
directories with suitable ownership to allow writing (and would probably
be possible to smuggle in a setuid-root shell if it was really necessary
to be root).