Make tasks wait for dependencies before running
Where a task has a strong dependency, the command is run through a
wrapper that first fetches an URL from the master controller, which
block until the dependencies have started.

The scheduler now has a Tornado-based HTTP server, which both serves the
wrapper script to the slaves, and responds to the requests to wait. This
unfortunately adds some complexity when used in a Docker container,
because the user has to provide the URL at which the server is reachable
from the outside world.