Implement the new master controller API under the hood
The katcp interface does not yet expose the new API. However, the stream
sources are passed to product_config.convert which generates a JSON doc
in the new schema, which is consumed by the graph generator.

The graph generator is moved into the katsdpcontroller package - because
it was completely rewritten and it was easier to do the rewrite in a
separate file, and because I never liked 'pip install katsdpcontroller'
installing things outside of the katsdpcontroller namespace. It is more
modular, with each node generated by a separate function.

The most user-visible change is that the names of the graph nodes has
changed: the sdp. prefix was removed (since it was used everywhere and
hence redundant) and the stream names are now used to form the node
names. This will affect the sensor names.