Make error messages from jsonschema more useful
The schema is adjusted to so that each input/output type is in its own
oneOf pair, with one branch matching anything not of that type, the
other giving the schema for that type. This makes the schema slightly
more verbose, but makes it possible to get clear error messages. The
error message priority function is also tweaked to ensure the error from
the correct branch is found.