Refactor configure and deconfigure
This is a first step towards adding new, alternative commands for
configuration and deconfiguration. The selection between showing status,
deconfiguration and configuring is hoisted up into
request_data_product_configure. A lot of the other work done in
configuration (either fresh or reconfigure) is also moved into a new
configure_product function and similarly for deconfiguration.