Handle capture-start/capture-stop as state transitions
This is made possible by the previous commit, which uses the strong
dependencies to order the requests in state transitions. Thus,
capture-start will be issued after the capture-inits, and capture-stop
before the capture-dones.

To make this work, the transition value was changed from just the
request name to the name plus arguments in a list.

This also makes the code to deal with simulators less hacky, since the
parameter to the request are stored inside the graph rather than parsed
out of the node name.