Improve unit test coverage of generator.py
This should prevent a repeat of the problem where a beamformer could not
be configured due to a NameError in the generator code. The new API is
used to configuring an all-singing all-dancing product that has all
three flavours of beamformer at the same time, to check that the
code-paths don't crash.