Fix for MKAIV-899 multicast addresses exhausted
Previously, a /16 network was allocated for SDP, and it was split into
/24 subnets, with a subnet for each stream name. Thus, for example, all
L0 streams in SDP went into a single subnet. This design was chosen at a
time when I thought that kernel routing tables would control multicast
interface binding. However, to manage network interfaces fully with
Mesos, it was necessary to do interface selection at the application
level.

There was also no recycling of addresses. This meant that, once ingest
was split into 4 nodes, only 32 subarrays could be built (or attempted
to be built) before running out of addresses.

The new design is a lot simpler: it allocates each subarray product a
/24 subnet for its needs (which should be far more than enough). When a
subarray product is terminated, the subnet is added back to a deque of
available subnets, so that there is no limit on how many subarrays can
be built as long as there are no more than 254 at a time.