Fix for Mesos master crashes
The symptoms are similar to MESOS-7197, but I think are caused by having
too many decimal places in a resource requirement even after Mesos
truncates to 3 places after the decimal point. This can happen for
bandwidths, which have lots of places before the decimal point. I'm not
sure exactly why this is a problem (I still need to investigate the
Mesos code and file a bug), but rounding the bandwidths to integers
seems to help.