Add --write-graphs option
This is a debugging option to visualise the various dependency
relationships. It could later be extended to also provide other
visualisations, such as the relationships between the streams in
the configuration dict.

This does add about 30MB to the Docker image by installing graphviz and
all its dependencies. It might be better to leave out support in the
Docker image and keep it solely for by-hand debugging.