Alternative approach to cleaning out dead subarray products
It was previously done by waiting on an asyncio event. That had some
issues (including failed unit tests) where a subarray product would go
to dead, and only in a later trip around the event loop would it go
away. The new approach uses a direct callback, which ensures that the
product goes to DEAD and gets cleaned out atomically.

There seem to be some issues with forced deconfiguration getting stuck,
but will fix that in a future request.