Rounding of output_channels to required alignment
This follows up on 340222e to add the functionality to correlator
ingest: any output_channels (that is non-empty and doesn't exceed the
range of channels) is now acceptable, and will be rounded up as
necessary to meet implementation requirements. This one is a little more
complicated because the rounding doesn't just affect the resource
requirements of the ingest process itself, but also anything that
consumes the stream. That's handled by modifying the config itself as
graph generation proceeds.

While writing the unit test, I also found a bug in _make_ingest: given a
continuum-only stream, it would be configured with a continuum_factor of 1
instead of the actual factor, due to a too-shallow copy.

_round_up was also broken.