Make ?product-reconfigure respect burndown
It will block until the old subarray is completely burned down before
configuring the new one.

There is a small race condition where between the old one being removed
and the new one being registered, another ?product-configure with the
same name could go through. Since reconfigure is intended only for
development work and CAM shouldn't be doing anything with a name that
might conflict, it's probably not worth trying to fix it.