Fix a potential race condition in product cleanup
The callback to remove the product from the product list when it died
was handled as a done callback on a future. However, setting a future
result only *schedules* the done callbacks for a future pass of the
event loop, rather than calling them immediately. This meant that other
requests could observe the dead product in the product list.

Fixed by going back to a dead event rather than a dead future (which is
simpler to wait for), plus a separate explicit list of callbacks.