More work in progress on Python 3 conversion
This is a start on a completely removal of tornado, replacing katcp with
aiokatcp and the tornado server with an aiohttp.web server. This
required adding a start method to Scheduler to bring up the web server,
since the name resolution happens asynchronously.

Currently the scheduler unit tests still fail.