Add scheduler support for multiple queues
Previously each launch group waited for tasks itself, rather than having
a central scheduler task.  Since there is now a need to arbitrate
between multiple queues, there is a central task (_launcher) which keeps
trying to fit a group and a set of resources.

The tricky part is cancellation: if a launch is cancelled, we need to
clean up. A difference from the previous version is that if a launch is
cancelled while it is being resolved, the resolution will be completed.