Run dummy postprocessing tasks
This is just to test out the plumbing - the actual tasks need to be
replaced by proper ones later.

It doesn't yet quite work as intended: telstate is killed off before the
postprocessing runs. We need some mechanism in the task class to wait
for all postprocessing to complete before shutting down telstate.