Use timestamps instead of counter to name capture blocks
This is as near as practical to guaranteeing that the CBID is globally
unique, without introducing long UUIDs or a separate database. I kept
the check that the CBID is unique amongst those known to the current
process (and also made that check global across subarray products),
since it was trivial to do.

To generate colliding IDs, it's now necessary to restart the master
controller and create a new capture block, with the same program block
ID, within the same second. Realistically it takes longer than that to
build a subarray, so it could only really happen if the clock was
adjusted too.