Use fixed-point decimal arithmetic for Mesos resources
This ensures that calculations will match the values computed by Mesos,
which rounded values to 3 decimal places after every computation. This
avoids some theoretical cases where the master controller might think
that some task or set of task just fits, but Mesos will decide that it
does not and fails.

The task nodes still use floating-point, and these values are converted
to Decimal when used in the scheduler's internal data structures.

Closes SR-760.