Pass simulated antenna positions through to cbfsim
This extends the configuration API to allow the positions of simulated
antennas to be passed in the 'simulated' key of the product config dict.
These are then passed on to cbfsim, allowing it to be fully configured
at startup.

For backwards compatibility, the old configure-subarray-from-telstate
mechanism is still in place, but eventually I'd like to remove it.