Add cal port for dask diagnostics
I'm busy switching cal to use dask.distributed, which provides a live
web UI. This plumbs it in by providing a port and a gui-url for it. It
should be safe even for existing cal (although the gui-url obviously
won't work).

I changed the order in which the task config is assembled so that
manually provided values override the automatics. This is needed so that
dask_diagnostics can be set to a host-port tuple instead of just a port.