Use a more appropriate redirect code
Redirect from an unknown (or not currently active) subarray was using
HTTP code 301, which is a permanent redirect and was being cached by
Chrome, making it stop working when the subarray comes back.

Now using 303 instead, which is a different sort of redirect and
isn't cached.

Also avoid redirecting favicon.ico so that it gets a 404 instead of
fetching /.