Add --s3-config-file option
This adds a config file (in JSON format) that specifies the S3 URL,
access key and secret key. It is currently used only to set
s3_endpoint_url for the vis_writer, but is available to the graph.

To keep argument lists from getting excessively long, I ripped out the
last remnants of --graph-override and the graph resolver (leftovers from
the old configuration API which did nothing).

The tests use open-mock-file (also known as open-file-mock: the PyPI
package name and the module name disagree), which I also rolled out to
the TestImageResolver tests to simplify them.