Use jinja2 to template product config schema
At the moment this adds code, but it should be more scalable for
updating the API as changes can be made just in the schema file without
having to add lots of special cases to the code. There is also the
possibility to add comments to the schema file in future (using Jinja2
comments), which hasn't been possible as JSON doesn't have comments, and
to simplify some of the constructs using macros.