Switch to use CaptureBlockState to represent transitions
This will make it possible to define other transitions. Since the state
machine for a capture block is linear, it also only needs a single state
to define a transition instead of a pair.

The call to configure-subarray-from-telstate was also removed - it
hasn't been needed for some time since kattelmod now passes the antenna
positions at product-configure time.