Postprocess transitions
This allows transitions to be defined to occur at the start of
postprocessing for a capture block. It is integrated with the watch on
capture-block-state sensors, so the steps for each process are:
1. Wait for dependencies to complete postprocessing.
2. Issue the transition commands, if any.
3. If there is a capture-block-state sensor, wait for the capture block
   to be finished according to that sensor.

The full meta write is now deferred to a postprocessing step, which
means it should capture the final outputs from cal.