Fixups to deconfigure shutdown ordering.
The mechanism introduced for keeping telstate alive until the end is
generalised and applied to meta_writer.

However, this mechanism is suboptimal. It should be possible to use the
existing graph information to kill things once they've cleared the
_exec_node_transition for going to state DEAD, without any separate
config.

Also cleaned up the various names assigned to cal in unit tests to
settle on 'cal'.