Track in-flight CBs per task
Each task now tracks the capture blocks for which capture-init was
issued (even if it didn't explicitly send any katcp request to it), but
for which the transition to state DEAD hasn't happened for that task.
This is integrated into _exec_node_transition so that each task can be
considered to be done with a CB as it finishes (whether that is
immediately, after a katcp request, or after a sensor update indicates
it is done).

Graceful shutdown is done by waiting for the task to no longer be ready
with any capture block.

The old wait_capture_blocks setting (which waits for all capture blocks
to be fully postprocessed) is kept for telstate, as it will eventually
be needed to keep telstate alive during batch processing, but for this
reason it is renamed to wait_batch.