Add support for Mesos maintenance windows
If an unavailability window is current or in the future, no new tasks
will be started on that machine. If it was in the past it is ignored, so
that a forgotten maintenance window won't permanently disable the box.
That's not entirely in line with Mesos recommendations (which seem to
imply that the times are purely advisory and the machine will be
DRAINING until taken down or the window cancelled), but is probably the
pragmatic thing to do.