Avoid reporting no resource offers if resolve() is slow
There is a corner case in launching if PhysicalTask.resolve() blocks for
a non-trivial amount of time. The launch group's future is waited for
under a wait_for, to ensure that a lack of resources causes a prompt
timeout. However, it can also fire if the resources have been acquired
but resolution is still in progress. This causes an unexpected "No
resources offers were received" message, as well as errors about the
launch failing.

A separate resources_future has been added, which is waited for with a
timeout, while the overall launch success is waited for without a
timeout. This also gives better protection against a timeout raised for
any other resource after the resources are acquired.

Closes SR-1093.