Add support for nvidia-container-runtime (aka nvidia-docker2)
It still gracefully falls back to the old method. The agent-mkconfig
script detects whether the Docker daemon has nvidia-container-runtime
registered, and sets an attribute
(katsdpcontroller.nvidia_container_runtime) if so. The scheduler then
passes the GPU UUID in the envar NVIDIA_VISIBLE_DEVICES, instead of
passing a driver volume mount and device passthrough.

The UUID field and katsdpcontroller.nvidia_container_runtime are
optional, so it will continue to operate with slaves that have not been
updated.

This also needs the images launched to be built using
ska-sa/katsdpdockerbase#10. That needs to be merged before using this in
anger.