Update to ICD v2.0
See
https://docs.google.com/document/d/11gBjHOTr8ijIwbdHMUe1qZ-c-HkGzYm63XIFGQJUPCA/edit
for the description of the API and what has changed in v2.0.

A major code change is that product_config now has a normalise function
that converts the config dictionary to a canonical form (in particular,
the latest version). This simplifies generator.py since it only has to
take one version into account. Normalisation also fills in defaults for
some optional values, which again simplifies generator.py.