Add support for sensors to specify their own Prom mappings
If the sensor description contains "prometheus: counter" or "prometheus:
gauge" the Prometheus metric will be dynamically generated and use
the description from the sensor (whichever one was first, which means
metric descriptions may change on each restart due to a race if multiple
services provide the same sensor with different descriptions).

The hard-coded list takes precedence, so for this to take effect we need
to add the hints to child services *and* then remove from the hardcoded
list.