Disable ibverbs for flag writer
It was originally turned on when we thought it was needed to keep up,
but that turned out to be down to switch configuration. It got turned
off for the vis writer, but the flag writer was overlooked.

This causes problems if cal and flag writer are on the same machine (as
occurs in the lab) because cal doesn't send with verbs, and so the
loopback doesn't work.