Remove Prometheus metrics when the subarray goes away
Previously it would linger but with the status metric set to
UNREACHABLE. Prometheus 1 recommended against removing metrics
completely because it didn't handle it very well (it would continue to
exist in the Prometheus database for 5 minutes). Prometheus 2 has better
handling (when it disappears in a scrape it is marked as gone in the
backend), so now it makes more sense to do this, and it should give more
accurate stats in the overview dashboard as dead subarrays will no
longer linger.