Use prometheus_async to serve Prometheus metrics
Instead of running a separate dedicated http server on port 8081 to
serve metrics, add a /metrics endpoint into the aiohttp server that is
used by the scheduler. The advantages are:

- One less port to deal with
- Metrics can only be served at yield points, which in particular means
  that the metrics representing katcp value and status are always in
  sync, whereas previously an update that changed value and status could
  be split across two scrapes.

In --interface mode the internal HTTP server is not started at all,
which suits CAM fine, so I've removed the --no-prometheus command-line
option (which was there for CAM's benefit to prevent it listening on a
port that collided).