Pass timeplot data via multicast
Apart from using multicast, it also passes it via the sdp_10g network,
which should reduce congestion on the 1Gb control network; and it
includes the bandwidth for the stream in the bandwidth estimates.

The matching support for katsdpdisp has already been merged. For ingest
it will require an extra --sdisp-interface option, but it should
actually work without it as long as we have the routing tables correctly
set up.