Major refactoring of resource management
This is needed to handle reservations. I suspect this has always been an
issue even before enabling multiple roles, but we haven't encountered it
because we haven't tried to use reservations. The issue is that when
some resources on an agent are reserved, Mesos won't coalesce reserved
and unreserved resources (or resources reserved to different roles), and
more importantly, the TaskInfo that the scheduler issues cannot coalesce
them either. So instead of tracking allocations simply as a number (for
scalars) or list of numbers (for ranges), we need to track the original
resource information.

Additionally, this takes care of allocating first from reserved
resources before unreserved (or less-specifically reserved, when there
is a hierarchy) resources, to avoid using up the unreserved resources
which could otherwise be offered to other roles.

While a lot of new code is introduced, in some places the code is
simplified because it no longer needs to distinguish between scalar and
range resources, and deals simply with the base class.

In most internal places, the attribute access (e.g. agent.cpus) has been
replaced by dictionaries (e.g. agent.resources['cpus']), which should
allow for better static type checking in future with tools like mypy.
The attribute access is retained in LogicalTask (via a class decorator
that sets descriptors) so that generator.py doesn't need to change. The
cores and ports mappings on PhysicalTask are similarly retained.