Support multiple S3 stores in s3_config.json
This is to support an on-site imaging buffer that is separate from the
transfer-to-CHPC buffer. To provide maximum flexibility I've provided
keys for "continuum", "spectral" and "archive" (the last being for
chunks that will eventually be archived).

This will need a bit of work during deployment to update the deployed
s3_config.json files.