Improve accuracy of signal display data rate calculation
The rate was being estimated as if there was only a single ingest, then
divided by the number of ingests - but this ignores that items like
bls_ordering and sd_timeseries don't get smaller when split across
ingests.

Make _timeplot_frame_size compute the per-ingest rate and scale up from
there when whole-system data rates are needed.

This addresses MKAIV-1269.