Update memory allocations for vis_writer/flag_writer
It hadn't been updated after the change to katsdpdatawriter, which uses
a deeper ringbuffer in vis_writer (probably because it took the code
from flag_writer) and also needs extra memory to handle the parallel
writing. The socket buffer calculation was also incorrect.

This will require about 55GB for 64A 32K. It could be reduced
substantially, either by reducing the safety margin (currently a factor
of 2) or by reducing the ringbuffer depth in katsdpdatawriter, but this
will get things going for now.

The number of workers is now also set explicitly to ensure that the
number used matches the number used for the calculation.