Wire in beamformer signal displays
The design is a little different to correlator signal displays, and I'm
not sure if I'm totally happy with it. In correlator sdisp, signal
displays are a byproduct of an SDP output (specifically, sdp_l0 type),
and the channel range shown is the output channel range. For beamformer,
we start a separate bf_ingest.py process for each input stream that
feeds at least one defined output, and it will always consume the full
bandwidth. That's pragmatic given that we might have engineering
beamformer, PTUSE beamformer, or even both, so we can't trivially
associate the signal display stream with a single SDP output.