Allow the whole of taskinfo to be overridden
Previously one could customise taskinfo.container in the LogicalTask,
but any other customisations to taskinfo had to be done by defining a
new PhysicalTask subclass and overriding resolve(). Now LogicalTask
allows the whole of taskinfo to be customised. This makes it easier to
customise things like taskinfo.command.environment or
taskinfo.command.shell.