Make framework_info.user a command-line option
This will allow the executor to be run as 'kat' on the agents, which
affects the sandbox ownership and hence will make the sandbox writable
for tasks running off docker-base with it's kat user (provided the UIDs
match!).

The sandbox is modified to pass in passwd and group files that contain a
kat user, and re-enable the switch-user functionality in the agent, so
that this can be tested.