Support multiple flags streams per cal
This uses the new --flags-streams option to cal (still on a branch) to
configure an arbitrary number of output flag streams with optional
continuum factor scaling.

It also fixes a bug in calculation of bandwidths for the flags stream,
which didn't take into account that flags are transmitted faster than
realtime.