Introduce an ERROR state to subarray products
At present it is triggered only by a failure to issue a request to a
child (e.g. if capture-init failed). Once a subarray is in this state,
it cannot be used with capture-init/capture-done, only deconfigured.
Once CAM detect this condition, it should avoid issues in future where
SDP gets into a semi-capturing state and can't go forward or backwards,
causing a whole prepared queue of capture blocks to all fail.

Future work is to extend it to also be triggered if one of the children
dies unexpectedly.

Relates to SR-1090.