Add a sensor for the currently configured products
Also took the opportunity to simplify some of the other sensor setup
code to set up a sensor with a single statement, rather than creating
the sensor, setting a value, and adding it. The time-synchronised sensor
is set to unknown rather than nominal, since we don't actually have an
interface to it.