Add continuum imager streams to sdp_archived_streams
While it's not archived in the CHPC archive, it's still archived, and it
needs to be in this list for katdal to autodetect the matching L1 flags
when loading the data.

Fixed by having `_make_vis_writer` and `_make_flag_writer` append to
sdp_archived_streams themselves, rather than the main generator code
having to remember to do it.