Integrate continuum imager pipeline
The integration still needs some work within the katsdpcontim image:
several logs are written to file and ideally should be integrated into
our logging system (tricky since it comes from Obit). At least the
Python logging should be done with katsdpservices.

Resource allocations are also a guess and will need to be tweaked.

The schema for s3_config.json is stricter and now requires read keys
(although they can be the same as the write keys). Production systems
will need to be updated for this.

Closes SR-1344.