Keep task state and mesos-state sensors alive longer
Record the sensors in the subarray product rather than the task, so that
they're only removed when the subarray product is removed, rather than
when the task dies. This fixes a bug in the bokeh dashboard where one
would get a KeyError if loading it at the wrong time, but also makes it
possible for other sensor consumers to keep tracking task state as tasks
die.

Also fixes some code that get mis-ordered in a faulty merge from master.