Disable datawriter dashboard
It turns out to be too slow for use in production. For engineering we
can use set-config-override to start it up, and we could also consider
adding a katcp command to the datawriter katcp server.

This depends on ska-sa/katsdpdata#133 being merged first: otherwise
every data writer will try to use the default port.