Show details of tasks when clicking on them
It is very ugly at the moment. It seems like Bokeh's DataTable insists
on having a fixed width and height rather than letting the browser sort
things out, so there is potentially a lot of unnecessary whitespace. The
tables within the task details also need a healthy sprinkling of CSS.