Run all capture-done transitions even if one fails
This partially addresses SR-1641: if some ingest fails capture-done, it
will nevertheless carry on to issue write-meta to write the light RDB
file so that isn't lost completely.

This unfortunately won't write the full RDB file. I'm hesitant to try to
work that in, because that waits until cal finishes with the capture
block and if we have failed transitions there is no guarantee that this
will ever happen.