Start on wiring in spectral imager
The logical graph for postprocessing is now constructed per capture
block (immediately before instantiating it), so that it can interrogate
the system for the targets to image (there are still some potential
issues if the catalogue contains multiple targets with the same name).
This requires pulling in katdal so that it can construct the same target
catalogue as the imager code sees.

At the moment there is no dependency between the spectral and continuum
imagers (and no continuum subtraction). There is little point in
providing that until the spectral imager can read L2 visibilities,
because otherwise continuum subtraction won't work as it'll be taking
the differences between differently-calibrated visibilities. Eventually
the scheduler module should take over launching these dependent but
separate batch tasks.