Fix unit tests
The introduction of katdal broke the unit tests badly (would gobble up
all the RAM) because telstate was a mock. It's now been replaced by a
real in-memory telstate, but that required further changes in the test
that were examining the mock calls to determine what had been added to
telstate.