Add support for non-realtime clocks in simulator
Adds `clock_ratio` and `start_time` options to `simulate` block in
product_config. Apart from passing it through to the simulator, it also
passes clock_ratio to ingest and cal so that they can scale their
transmission rates to keep up (PRs to follow).

At the moment using start_time makes it impractical to do more than one
capture block in the lifetime of the subarray, because they will all
happen at the same time. Fixing that will need interface changes to
`capture-init` to communicate the per-CB start time.

These are rolled into the 2.4 version of the ICD.