Run metawriter after postprocessing tasks
To make this a general solution, an extra state is added to capture
blocks. They now first go through BURNDOWN (where cal and flag writer
wind things up) and then POSTPROCESSING (imaging tasks). This provides
finer granularity in specifying transition points.

A nice side benefit is that the deconfigure_wait and
wait_capture_blocks_dead flags can be replaced by an indication of how
long each node is involved in the life of each capture block, which in
turn controls at what point they may be killed. Most die on the
transition to BURNDOWN, cal and flag writer on the transition to
POSTPROCESSING, and metawriter and telstate only at the end (on the
transition to DEAD).