Run separate continuum imager process per target
This gets a bit complicated, because the katdal catalogue creation is
sensitive (in corner cases) to which L0 stream is used. We use a cache
keyed by the underlying baseline-correlation-products to ensure that we
get consistent results between the continuum and spectral imagers. We
then also have to deal with the case that the continuum image won't see
that target, in which case we skip both it and the downstream spectral
imaging.