Fix use of the wrong stream name
There is (for example) sdp_l0_continuum (defined in the config) and
continuum_image.sdp_l0_continuum (the data written to the imager buffer
and with an associated chunk_info). There might also be a chunk_info for
sdp_l0_continuum if it is being archived, but not necessarily. So we
need to pass continuum_image.sdp_l0_continuum to katdal when looking up
targets.