Pass --outputdir to continuum imager
This gives it a place to write images that will be picked up by
katsdptransfer.

There is also some code that tracks the number of times a physical task
has been retried, to simplify generating names that are unique across
all physical tasks, but it ended up not being needed because
katsdpcontim now generates a unique directory name itself.