React faster to a product dying
When the product controller shuts down, it emits a #disconnect inform.
We now respond to that to immediately mark the product as dead and
remove it from the product list. Saving to Zookeeper is done in the
background. In some cases this will mean double saves to Zookeeper, but
that is harmless.