Make everything work again
The sandbox docker-compose got another revamp: my approach of trying to
hardcode 172.17.0.1 as the bridge address didn't work, because the
address can vary. Now the product controller is run with host
networking, looking at the environment variables PORTn to set its ports.

Singularity is also bound to 127.0.0.1 to avoid exposing the host
it to the world (although the tasks launched via Docker are all
world-exposed).